@@ -1,17 +1,13 @@
 package com.yhao.floatwindow;
 
 import android.content.Context;
-import android.graphics.PixelFormat;
 import android.graphics.Point;
 import android.graphics.Rect;
-import android.os.Build;
-import android.provider.Settings;
-import android.support.annotation.RequiresApi;
+
+import androidx.annotation.RequiresApi;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.WindowManager;
-
-import java.lang.reflect.Method;
 
 /**
  * Created by yhao on 2017/12/22.
